[#3869] Change to the structure that can change codec processing in rpc
Change to the structure that can change codec processing in rpc
