+++ collector/src/main/java/com/navercorp/pinpoint/collector/receiver/tcp/AgentBaseDataReceiver.java
@@ -18,9 +18,7 @@ package com.navercorp.pinpoint.collector.receiver.tcp;
 
 import com.navercorp.pinpoint.collector.cluster.zookeeper.ZookeeperClusterService;
 import com.navercorp.pinpoint.collector.config.AgentBaseDataReceiverConfiguration;
-import com.navercorp.pinpoint.collector.receiver.AddressFilterAdaptor;
 import com.navercorp.pinpoint.collector.receiver.DispatchHandler;
-import com.navercorp.pinpoint.common.server.util.AddressFilter;
 import com.navercorp.pinpoint.collector.rpc.handler.AgentLifeCycleHandler;
 import com.navercorp.pinpoint.collector.service.AgentEventService;
 import com.navercorp.pinpoint.common.server.util.AgentEventType;
@@ -33,7 +31,6 @@ import com.navercorp.pinpoint.rpc.packet.HandshakeResponseType;
 import com.navercorp.pinpoint.rpc.packet.PingPayloadPacket;
 import com.navercorp.pinpoint.rpc.packet.RequestPacket;
 import com.navercorp.pinpoint.rpc.packet.SendPacket;
-import com.navercorp.pinpoint.rpc.server.ChannelFilter;
 import com.navercorp.pinpoint.rpc.server.PinpointServer;
 import com.navercorp.pinpoint.rpc.server.PinpointServerAcceptor;
 import com.navercorp.pinpoint.rpc.server.ServerMessageListener;
@@ -59,10 +56,9 @@ public class AgentBaseDataReceiver {
 
     private final Logger logger = LoggerFactory.getLogger(AgentBaseDataReceiver.class);
 
-    private PinpointServerAcceptor serverAcceptor;
+    private PinpointServerAcceptor acceptor;
 
     private final AgentBaseDataReceiverConfiguration configuration;
-    private final AddressFilter addressFilter;
 
     private final ZookeeperClusterService clusterService;
 
@@ -80,14 +76,14 @@ public class AgentBaseDataReceiver {
     @Resource(name = "channelStateChangeEventHandlers")
     private List<ServerStateChangeEventHandler> channelStateChangeEventHandlers = Collections.emptyList();
 
-    public AgentBaseDataReceiver(AgentBaseDataReceiverConfiguration configuration, Executor executor, AddressFilter addressFilter, DispatchHandler dispatchHandler) {
-        this(configuration, executor, addressFilter, dispatchHandler, null);
+    public AgentBaseDataReceiver(AgentBaseDataReceiverConfiguration configuration, Executor executor, PinpointServerAcceptor acceptor, DispatchHandler dispatchHandler) {
+        this(configuration, executor, acceptor, dispatchHandler, null);
     }
 
-    public AgentBaseDataReceiver(AgentBaseDataReceiverConfiguration configuration, Executor executor, AddressFilter addressFilter, DispatchHandler dispatchHandler, ZookeeperClusterService service) {
+    public AgentBaseDataReceiver(AgentBaseDataReceiverConfiguration configuration, Executor executor, PinpointServerAcceptor acceptor, DispatchHandler dispatchHandler, ZookeeperClusterService service) {
         this.configuration = Assert.requireNonNull(configuration, "config must not be null");
         this.executor = Objects.requireNonNull(executor, "executor must not be null");
-        this.addressFilter = Objects.requireNonNull(addressFilter, "addressFilter must not be null");
+        this.acceptor = Objects.requireNonNull(acceptor, "acceptor must not be null");
 
         this.tcpPacketHandler = wrapDispatchHandler(dispatchHandler);
         this.clusterService = service;
@@ -105,8 +101,7 @@ public class AgentBaseDataReceiver {
         if (logger.isInfoEnabled()) {
             logger.info("start() started");
         }
-        ChannelFilter connectedFilter = new AddressFilterAdaptor(addressFilter);
-        PinpointServerAcceptor acceptor = new PinpointServerAcceptor(connectedFilter);
+
         prepare(acceptor);
 
         // take care when attaching message handlers as events are generated from the IO thread.
@@ -148,8 +143,6 @@ public class AgentBaseDataReceiver {
             }
         });
         acceptor.bind(configuration.getBindIp(), configuration.getBindPort());
-
-        this.serverAcceptor = acceptor;
 
         if (logger.isInfoEnabled()) {
             logger.info("start() completed");
@@ -203,10 +196,9 @@ public class AgentBaseDataReceiver {
             logger.info("stop() started");
         }
 
-        if (serverAcceptor != null) {
-            serverAcceptor.close();
+        if (acceptor != null) {
+            acceptor.close();
         }
-
 
         if (logger.isInfoEnabled()) {
             logger.info("stop() completed");
