+++ rpc/src/main/java/com/navercorp/pinpoint/rpc/PipelineFactory.java
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,14 +14,15 @@
  * limitations under the License.
  */
 
-package com.navercorp.pinpoint.rpc.client;
+package com.navercorp.pinpoint.rpc;
 
-import org.jboss.netty.util.Timer;
+import org.jboss.netty.channel.ChannelPipeline;
 
 /**
- * @author Woonduk Kang(emeroad)
+ * @author Taejin Koo
  */
-public interface ClientHandlerFactory {
+public interface PipelineFactory {
 
-    PinpointClientHandler newClientHandler(ConnectionFactory connectionFactory, SocketAddressProvider remoteAddressProvider, Timer timer, boolean reconnect);
+    ChannelPipeline newPipeline();
+
 }
