+++ rpc/src/main/java/com/navercorp/pinpoint/rpc/server/ServerCodecPipelineFactory.java
@@ -1,0 +1,43 @@
+/*
+ * Copyright 2018 NAVER Corp.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.navercorp.pinpoint.rpc.server;
+
+import com.navercorp.pinpoint.rpc.PipelineFactory;
+import com.navercorp.pinpoint.rpc.codec.PacketEncoder;
+import com.navercorp.pinpoint.rpc.codec.ServerPacketDecoder;
+import org.jboss.netty.channel.ChannelPipeline;
+import org.jboss.netty.channel.Channels;
+
+/**
+ * @author Taejin Koo
+ */
+public class ServerCodecPipelineFactory implements PipelineFactory {
+
+    public ServerCodecPipelineFactory() {
+    }
+
+    @Override
+    public ChannelPipeline newPipeline() {
+        ChannelPipeline pipeline = Channels.pipeline();
+
+        pipeline.addLast("decoder", new ServerPacketDecoder());
+        pipeline.addLast("encoder", new PacketEncoder());
+
+        return pipeline;
+    }
+
+}
